# Conflicts:
#	backend/src/main/java/com/stockmatch/config/SecurityConfig.java

+++ backend/src/main/java/com/stockmatch/portfolio/service/PortfolioValuationService.java
@@ -352,7 +352,7 @@
             var price = security.isKorean()
                     ? stockPriceService.getKrStockPrice(ticker).getCurrentPrice()
                     : stockPriceService.getUsStockPrice(ticker).getCurrentPrice();
-            return BigDecimal.valueOf(price);
+            return price;
         }
 
         /**
